+++ src/test/java/io/github/astrapi69/evaluate/object/evaluator/ToStringEvaluatorTest.java
@@ -23,12 +23,12 @@
  */
 package io.github.astrapi69.evaluate.object.evaluator;
 
-import static org.testng.AssertJUnit.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.Serializable;
 
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
 
 import io.github.astrapi69.evaluate.object.Person;
 import io.github.astrapi69.random.object.RandomStringFactory;