+++ src/test/java/io/github/astrapi69/evaluate/object/verifier/ContractVerifierTest.java
@@ -20,7 +20,7 @@
  */
 package io.github.astrapi69.evaluate.object.verifier;
 
-import org.testng.annotations.Test;
+import org.junit.jupiter.api.Test;
 
 import io.github.astrapi69.test.object.A;
 import io.github.astrapi69.test.object.AlgorithmModel;
@@ -43,7 +43,6 @@ import io.github.astrapi69.test.object.annotation.classtype.SubAnnotatedClass;
 import io.github.astrapi69.test.object.auth.AccessRight;
 import io.github.astrapi69.test.object.auth.Role;
 import io.github.astrapi69.test.object.auth.Roles;
-import io.github.astrapi69.test.object.exception.ExceptionEvent;
 
 /**
  * The unit test class for the class {@link ContractVerifier}